requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-modeltranslation==0.15.2
django-rosetta==0.9.4
django-storages==1.10
django-taggit==1.3.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
polib==1.1.0
psycopg2-binary==2.8.5
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
